Show correct Assembly version (#346)
When Umbra is in testing, the version is displayed incorrectly because
it uses `AssemblyVersion` from the manifest instead of checking
`IDalamudPluginInterface.IsTesting` and using `TestingAssemblyVersion`.

But there is a different solution to this, by using the version from the
executing assembly directly.
I've never seen it being null, but added a check nonetheless.