Hide looted Treasure Coffer Markers (#356)
When you loot a Treasure Chest and walk away, the marker is still
visible until the GameObject despawns.
This change makes it so the marker is removed as soon as the treasure is
fading out after being looted, while still being shown when loot is
pending.

I wasn't sure if the marker should be removed when the chest is opened
or fading out, so I only opted in for the fading out condition.

See
https://github.com/aers/FFXIVClientStructs/blob/main/FFXIVClientStructs/FFXIV/Client/Game/Object/Treasure.cs