adjust pot fate respawn timer logic (#367)
When I adjusted the Bunny FATE Markers for the OC Pot FATEs, the
assumption was that the respawn logic behaved the same.
So anyway turns out that was a forking lie.

Pot FATEs respawn on a set 30 minute cycle after the last one spawned.
The kill time aswell as the Pots being alive after have nothing to do
with this.

## To-do
- [x] Adjust the time and text in the tooltip since it's no longer a
range
- [x] Make it start counting from FATE spawn rather than FATE last seen